[refactor] Embed User as field, some other refactors - BotPrediction and Report now have User as EmbeddedDocument - BotPrediction schema is now UserInfo schema - rename db collection name - rename tweet.py to twitter.py to contain both tweet and user - fix broken tests, change db query to match new model structure
