+++ app/models/__init__.py
@@ -1,4 +1,4 @@
 from .bot_prediction import BotPrediction
 from .operator import Operator
 from .report import Report
-from .tweet import Tweet
+from .twitter import Tweet, User
