+++ app/models/report.py
@@ -9,14 +9,13 @@ from mongoengine import (
     EmbeddedDocumentField,
     EmbeddedDocumentListField,
     FloatField,
-    IntField,
     ListField,
     StringField,
 )
 
 from app.schemas import ReportResponse
 
-from .tweet import Tweet
+from .twitter import Tweet, User
 
 
 class ReportKey(EmbeddedDocument):
@@ -26,19 +25,10 @@ class ReportKey(EmbeddedDocument):
 
 class Report(Document):
     report_key = EmbeddedDocumentField(ReportKey, primary_key=True)
-    tweets_count: int = IntField(required=True)
-    name: str = StringField(required=True)
-    username: str = StringField(required=True)
-    created_at: datetime = DateTimeField(required=True)
-    followers_count: int = IntField(required=True)
-    followings_count: int = IntField(required=True)
-    favourites_count: int = IntField(required=True)
-    listed_count: int = IntField(required=True)
-    default_profile: bool = BooleanField(required=True)
-    default_profile_image: bool = BooleanField(required=True)
-    protected: bool = BooleanField(required=True)
-    verified: bool = BooleanField(required=True)
-    avatar: str = StringField(required=True)
+
+    # NOTE: User.banner is never used in report!
+    user = EmbeddedDocumentField(User, required=True)
+
     tweets: List[Tweet] = EmbeddedDocumentListField(Tweet, default=[])
     reporters: List[str] = ListField(StringField(), required=True)
     score: float = FloatField(required=True)
@@ -49,9 +39,9 @@ class Report(Document):
     def to_response(self) -> ReportResponse:
         response = ReportResponse(
             id=self.report_key.twitter_id,
-            avatar=self.avatar,
-            username=self.username,
-            created_at=self.created_at,
+            avatar=self.user.avatar,
+            username=self.user.username,
+            created_at=self.user.created_at,
             scrape_date=self.report_key.scrape_date,
             report_count=len(self.reporters),
             score=self.score,
