+++ app/schemas/__init__.py
@@ -1,17 +1,16 @@
 from pydantic import BaseModel
 
 from .auth import AccessToken, AuthDetails, TokenPayload, UserIdentifier
-from .bot_prediction import BotPrediction
 from .report import ReportResponse
-from .tweet import TweetInfo, TweetResponse
+from .twitter import TweetInfo, TweetResponse, UserInfo
 
 
 class CheckResponse(BaseModel):
     score: float
-    user_info: BotPrediction
+    user_info: UserInfo
 
 
 class DetailResponse(BaseModel):
-    user_info: BotPrediction
+    user_info: UserInfo
     tweet_info: TweetInfo
     score: float
