+++ app/schemas/twitter.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import List
+from typing import List, Optional
 
 from pydantic import BaseModel
 
@@ -22,3 +22,15 @@ class TweetStatistics(BaseModel):
 
 class TweetInfo(TweetStatistics):
     recent_tweets: List[TweetResponse]
+
+
+class UserInfo(BaseModel):
+    id: str
+    name: str
+    username: str
+    created_at: datetime
+    followers_count: int
+    followings_count: int
+    verified: bool
+    avatar: str
+    banner: Optional[str]
